rename test package names
