+++ serialization/src/test/resources/rune-serializer-round-trip-test/basic/basic-types-single.json
@@ -1,6 +1,6 @@
 {
-  "@model" : "test",
-  "@type" : "test.basic.Root",
+  "@model" : "serialization",
+  "@type" : "serialization.test.basic.Root",
   "@version" : "0.0.0",
   "basicSingle" : {
     "booleanType" : true,
