+++ serialization/src/test/resources/rune-serializer-round-trip-test/basic/basic.rosetta
@@ -1,4 +1,4 @@
-namespace test.basic
+namespace serialization.test.basic
 
 annotation rootType: <"Mark a type as a root of the rosetta model">
