+++ serialization/src/test/resources/rune-serializer-round-trip-test/choicetype/choice-mixed-2.json
@@ -1,6 +1,6 @@
 {
-  "@model" : "test",
-  "@type" : "test.choicetype.Root",
+  "@model" : "serialization",
+  "@type" : "serialization.test.choicetype.Root",
   "@version" : "0.0.0",
   "choiceMixed" : {
     "string" : "foo"
