+++ serialization/src/test/resources/rune-serializer-round-trip-test/data/data-types.json
@@ -1,6 +1,6 @@
 {
-  "@model" : "test",
-  "@type" : "test.data.Root",
+  "@model" : "serialization",
+  "@type" : "serialization.test.data.Root",
   "@version" : "0.0.0",
   "a" : {
     "b" : [ { }, {
