+++ serialization/src/test/resources/rune-serializer-round-trip-test/data/data.rosetta
@@ -1,4 +1,4 @@
-namespace test.data
+namespace serialization.test.data
 
 annotation rootType: <"Mark a type as a root of the rosetta model">
