+++ serialization/src/test/resources/rune-serializer-round-trip-test/enumtypes/enum-types-list.json
@@ -1,6 +1,6 @@
 {
-  "@model" : "test",
-  "@type" : "test.enumtypes.Root",
+  "@model" : "serialization",
+  "@type" : "serialization.test.enumtypes.Root",
   "@version" : "0.0.0",
   "enumList" : {
     "enumType" : [ "A", "B", "C", "B" ]
