+++ serialization/src/test/resources/rune-serializer-round-trip-test/enumtypes/enum.rosetta
@@ -1,4 +1,4 @@
-namespace test.enumtypes
+namespace serialization.test.enumtypes
 
 annotation rootType: <"Mark a type as a root of the rosetta model">
