+++ serialization/src/test/resources/rune-serializer-round-trip-test/extension/base-type.json
@@ -1,6 +1,6 @@
 {
-  "@model" : "test",
-  "@type" : "test.extension.Root",
+  "@model" : "serialization",
+  "@type" : "serialization.test.extension.Root",
   "@version" : "0.0.0",
   "typeA" : {
     "fieldA" : "foo"
