+++ serialization/src/test/resources/rune-serializer-round-trip-test/extension/extended-type-polymorphic.json
@@ -1,9 +1,9 @@
 {
-  "@model" : "test",
-  "@type" : "test.extension.Root",
+  "@model" : "serialization",
+  "@type" : "serialization.test.extension.Root",
   "@version" : "0.0.0",
   "typeA" : {
-    "@type" : "test.extension.B",
+    "@type" : "serialization.test.extension.B",
     "fieldA" : "bar",
     "fieldB" : "foo"
   }
