+++ serialization/src/test/resources/rune-serializer-round-trip-test/metalocation/address.json
@@ -1,6 +1,6 @@
 {
-  "@model" : "test",
-  "@type" : "test.metalocation.Root",
+  "@model" : "serialization",
+  "@type" : "serialization.test.metalocation.Root",
   "@version" : "0.0.0",
   "bAddress" : {
     "@ref:scoped" : "someLocation"
