+++ serialization/src/test/resources/rune-serializer-round-trip-test/metalocation/meta-location.rosetta
@@ -1,4 +1,4 @@
-namespace test.metalocation
+namespace serialization.test.metalocation
 
 annotation rootType: <"Mark a type as a root of the rosetta model">
