+++ serialization/src/test/resources/rune-serializer-round-trip-test/metascheme/meta-scheme.rosetta
@@ -1,4 +1,4 @@
-namespace test.metascheme
+namespace serialization.test.metascheme
 
 annotation rootType: <"Mark a type as a root of the rosetta model">
