+++ serialization/src/test/resources/rune-serializer-round-trip-test/record/record-types-single.json
@@ -1,6 +1,6 @@
 {
-  "@model" : "test",
-  "@type" : "test.record.Root",
+  "@model" : "serialization",
+  "@type" : "serialization.test.record.Root",
   "@version" : "0.0.0",
   "recordSingle" : {
     "dateTimeType" : "2024-12-12T11:18:58",
