+++ serialization/src/test/resources/rune-serializer-round-trip-test/record/record.rosetta
@@ -1,4 +1,4 @@
-namespace test.record
+namespace serialization.test.record
 
 annotation rootType: <"Mark a type as a root of the rosetta model">
